requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
debugpy==1.5.1
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.0.6
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==19.0.2
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
setuptools==66.0.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torch-tb-profiler==0.4.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.4
wheel==0.38.4
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
